requirements.txt
fastapi
uvicorn
sqlmodel
python-jose
passlib[bcrypt]
python-dotenv
streamlit
httpx
plotly